MAINT: rename read_edf.py module to read.py
This is so that we don't confused the read_edf.py module with the read_edf function that
is part of the read_edf.py module.

The documented public API was and remains `from eyelinkio import read_edf` , which imports the function,
so this should not have any compat issues with existing user code, unless they explicitly are doing something
like eyelinkio.edf.read_edf.some_function
