+++ eyelinkio/edf/__init__.py
@@ -1,2 +1,2 @@
 # License: BSD (3-clause)
-from .read_edf import read_edf, EDF
+from .read import read_edf, EDF
